add solution logic to color-render challenge with minimal css styling
