--- src/code_challenges/color_renderer/Color.tsx
@@ -1,10 +1,0 @@
-
-export default function Color ({ hex, name }) {
-    return (
-      <div
-        className='color-square' style={{ backgroundColor: hex }}
-      >
-        <h2>{name}</h2>
-      </div>
-    )
-  }
