+++ src/components/code_challenges/color_renderer/Color.tsx
@@ -1,0 +1,16 @@
+import './Color.css'
+
+interface Props{
+    hex:string
+    name:string
+}
+
+export default function Color ({ hex, name }:Props){
+    return (
+      <div
+        className='color-square' style={{ backgroundColor: hex }}
+      >
+      <h3>{name}</h3>
+      </div>
+    )
+  }
