+++ src/components/code_challenges/color_renderer/ColorRenderer.css
@@ -1,0 +1,4 @@
+.color-rederer-wrapper{
+    border: black solid 3px;
+    border-radius: 5px;
+}
